Extract bundle_election module in sp-domains
